modified:   src/components/Elements/Button.jsx         renamed:    src/components/Fragments/FooterMenu.jsx -> src/components/Elements/FooterMenu.jsx         modified:   src/components/Fragments/Footer.jsx         modified:   src/components/Fragments/Navbar.jsx         modified:   src/page/Login.jsx         modified:   src/routes/Routes.jsx         modified:   src/style/detail-user.css
